only owners can delete and update tags, fix on delete cascade for tags

+++ src/tag/tag.controller.ts
@@ -12,18 +12,24 @@ import {
   ParseIntPipe,
   Put,
   NotFoundException,
+  ForbiddenException,
 } from '@nestjs/common';
 import { TagService } from './tag.service';
 import { CreateTagDto } from './dto/create-tag.dto';
 import { UpdateTagDto } from './dto/update-tag.dto';
 import { GetTagsPrams } from './types/get-tags.params';
-import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
+import { JwtAuthGuard } from '../auth/jwt-auth.guard';
 import { TagsResponseBody } from './types/tags-response-body.type';
+import { CaslAbilityFactory } from '../casl/casl-ability.factory';
+import { Action } from '../casl/action.enum';
 
 @UseGuards(JwtAuthGuard)
 @Controller('tag')
 export class TagController {
-  constructor(private readonly tagService: TagService) {}
+  constructor(
+    private readonly tagService: TagService,
+    private readonly caslAbilityFactory: CaslAbilityFactory,
+  ) {}
 
   @Post()
   async create(@Request() req, @Body() createTagDto: CreateTagDto) {
@@ -66,23 +72,39 @@ export class TagController {
 
   @Put(':id')
   async update(
+    @Request() req,
     @Param('id', ParseIntPipe) id: number,
     @Body() updateTagDto: UpdateTagDto,
   ) {
-    const result = await this.tagService.update(id, updateTagDto);
+    const ability = this.caslAbilityFactory.createForUser(req.user);
+    const tag = await this.findOne(id);
 
-    if (!result.affected) {
+    if (!tag) {
       throw new NotFoundException();
+    }
+
+    if (ability.can(Action.Update, tag)) {
+      await this.tagService.update(id, updateTagDto);
+    } else {
+      throw new ForbiddenException('You must be the owner of the tag');
     }
 
     return this.tagService.findOne(id);
   }
 
   @Delete(':id')
-  async remove(@Param('id', ParseIntPipe) id: number) {
-    const result = await this.tagService.remove(id);
-    if (!result.affected) {
+  async remove(@Request() req, @Param('id', ParseIntPipe) id: number) {
+    const ability = this.caslAbilityFactory.createForUser(req.user);
+    const tag = await this.findOne(id);
+
+    if (!tag) {
       throw new NotFoundException();
+    }
+
+    if (ability.can(Action.Delete, tag)) {
+      this.tagService.remove(id);
+    } else {
+      throw new ForbiddenException('You must be the owner of the tag');
     }
   }
 }
